requirements.txt
pytesseract
pandas
pdf2image
pymupdf
pdfplumber
python-Levenshtein
dkpro-cassis
requests